requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.5
Django==4.1
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-writable-nested==0.7.0
filelock==3.8.0
idna==3.4
Pillow==9.2.0
platformdirs==2.5.2
PyJWT==2.5.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.3
